requirements.txt
flask==3.0.2
python-dotenv==1.0.1
supabase==2.3.4
pytest==8.0.2
black==24.2.0
flake8==7.0.0 